BL-72 <created searchbar folder in table and moved files there for better organization>
